3_semestre/IDP-BD-2025.1/IDP-Mensager/app/requirements.txt
flask
pymongo
pika
python-dotenv
flask-socketio
